clear info msgs

+++ Chess98/board.hpp
@@ -649,14 +649,14 @@ int Board::knightMobility() const
         {
             if (this->teamOn(x - 2, y - 1) != this->team && this->teamOn(x - 2, y - 1) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x - 2][y - 1] != 1)
+                if (badKnightPosMap[size_t(x - 2)][size_t(y - 1)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
             }
             if (this->teamOn(x - 2, y + 1) != this->team && this->teamOn(x - 2, y + 1) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x - 2][y + 1] != 1)
+                if (badKnightPosMap[size_t(x - 2)][size_t(y + 1)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
@@ -666,14 +666,14 @@ int Board::knightMobility() const
         {
             if (this->teamOn(x + 2, y - 1) != this->team && this->teamOn(x + 2, y - 1) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x + 2][y - 1] != 1)
+                if (badKnightPosMap[size_t(x + 2)][size_t(y - 1)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
             }
             if (this->teamOn(x + 2, y + 1) != this->team && this->teamOn(x + 2, y + 1) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x + 2][y + 1] != 1)
+                if (badKnightPosMap[size_t(x + 2)][size_t(y + 1)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
@@ -683,14 +683,14 @@ int Board::knightMobility() const
         {
             if (this->teamOn(x - 1, y - 2) != this->team && this->teamOn(x - 1, y - 2) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x - 1][y - 2] != 1)
+                if (badKnightPosMap[size_t(x - 1)][size_t(y - 2)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
             }
             if (this->teamOn(x + 1, y - 2) != this->team && this->teamOn(x + 1, y - 2) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x + 1][y - 2] != 1)
+                if (badKnightPosMap[size_t(x + 1)][size_t(y - 2)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
@@ -700,14 +700,14 @@ int Board::knightMobility() const
         {
             if (this->teamOn(x - 1, y + 2) != this->team && this->teamOn(x - 1, y + 2) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x - 1][y + 2] != 1)
+                if (badKnightPosMap[size_t(x - 1)][size_t(y + 2)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
             }
             if (this->teamOn(x + 1, y + 2) != this->team && this->teamOn(x + 1, y + 2) != OVERFLOW_TEAM)
             {
-                if (badKnightPosMap[x + 1][y + 2] != 1)
+                if (badKnightPosMap[size_t(x + 1)][size_t(y + 2)] != 1)
                 {
                     result += KNIGHT_EXTEND;
                 }
